fixed bugs

+++ src/staggered_grid.cpp
@@ -1,7 +1,7 @@
 #include "../includes/staggered_grid.h"
 
 
-StaggeredGrid::StaggeredGrid(const std::array<int,2> size,const std::array<int,2> physicalSize)):
+StaggeredGrid::StaggeredGrid(const std::array<int,2> size,const std::array<double,2> physicalSize):
 u_(size, VRIGHT, physicalSize), v_(size, VTOP, physicalSize), p_(size, VCENTRE, physicalSize), f_(size, VRIGHT, physicalSize), g_(size, VTOP, physicalSize), rhs_(size, VCENTRE, physicalSize)
 {}
